pyproject.toml
[tool.poetry]
name = "email_assistant"
version = "0.1.0"
description = "AI Email Assistant for Executive Productivity"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
flask = "^3.1.1"
crewai = "^0.126.0"
google-api-python-client = "^2.172.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
firebase-admin = "^6.9.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"